fix: properly filter out non-composed events dispatched on the shadow root
